+++ src/backend/server.ts
@@ -5,22 +5,16 @@ import express, { type Express } from 'express';
 import helmet from 'helmet';
 import cors from 'cors';
 import morgan from 'morgan';
-import cookieParser from 'cookie-parser';
 
 import { Handler } from '@/api/handler';
 import { EchoHandler } from '@/api/echo';
 import { ServeAppHandler } from '@/api/serveApp';
-import { GenerateNonceHandler } from '@/api/auth/getNonce';
-import { CreateAccountHandler } from '@/api/auth/createAccount';
-import { GetAuthTokenHandler } from '@/api/auth/getAuthToken';
 import { PostgresDB } from '@/database/postgres';
 import { errorHandler } from '@/middleware/errorHandler';
-import { authNHandler } from '@/middleware/authNHandler';
+import { authMiddleware } from '@/middleware/authMiddleware';
 import { CreateGameHandler } from '@/api/game/createGame';
 import { JoinGameHandler } from '@/api/game/joinGame';
-import { EventHandler } from '@/api/event';
 import { GetGameStateHandler } from '@/api/game/getGameState';
-import { BatchGetUsernameHandler } from '@/api/auth/batchGetUsername';
 
 export class Server {
     private readonly app: Express;
@@ -31,7 +25,6 @@ export class Server {
         // add middleware
         this.app.use(express.json());
         this.app.use(helmet());
-        this.app.use(cookieParser())
         this.app.use(cors({origin: 'http://frontend:80'}));
         this.app.use(morgan(':method :url', {immediate: true}));
         this.app.use(morgan(':method :url :status :res[content-length] - :response-time ms'));
@@ -40,21 +33,15 @@ export class Server {
         new Map<string, Handler>([
             ["/", ServeAppHandler.INSTANCE],
             ["/echo", EchoHandler.INSTANCE],
-            ["/auth/getNonce", GenerateNonceHandler.INSTANCE],
-            ["/auth/createAccount", CreateAccountHandler.INSTANCE],
-            ["/auth/getAuthToken", GetAuthTokenHandler.INSTANCE],
-            ["/auth/batchGetUsername", BatchGetUsernameHandler.INSTANCE],
         ]).forEach((handler: Handler, path: string) => {
             this.app.use(path, handler.router);
         });
         new Map<string, Handler>([
-            ["/authNedEcho", EchoHandler.INSTANCE],
-            ["/event", EventHandler.INSTANCE],
             ["/createGame", CreateGameHandler.INSTANCE],
             ["/joinGame", JoinGameHandler.INSTANCE],
             ["/getGameState", GetGameStateHandler.INSTANCE],
         ]).forEach((handler: Handler, path: string) => {
-            this.app.use(path, authNHandler, handler.router);
+            this.app.use(path, authMiddleware, handler.router);
         });
         // register error middleware
         this.app.use(errorHandler);
@@ -66,7 +53,7 @@ export class Server {
     public async start() {
         // initialize database
         await PostgresDB.INSTANCE.initialize();
-        
+
         // start server
         const port = process.env.BACKEND_PORT || 3000;
         console.log(`Listening on port ${port}`);