+++ src/frontend/component/LoginView.vue
@@ -1,11 +1,10 @@
 <script setup lang="ts">
 import { defineProps, ref } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
-import { AuthNService } from '@/service/authNService';
-import { EventSourceSingleton } from '@/util/sse';
+import { signIn } from '@/service/authService';
 
 const authModel = defineProps({
-    username: { type: String, required: true },
+    email: { type: String, required: true },
     password: { type: String, required: true },
 });
 
@@ -15,10 +14,7 @@ const route = useRoute();
 const router = useRouter();
 
 async function sendLogin() {
-    AuthNService.login(authModel.username, authModel.password).then((_jwt: string) => {;
-        // open EventSource connection if not already
-        EventSourceSingleton.getInstance();
-
+    signIn(authModel.email, authModel.password).then(() => {
         // redirect to previous page
         const redirectedFrom = route.query.redirect;
         if (redirectedFrom !== undefined && typeof redirectedFrom === "string") {
@@ -31,7 +27,7 @@ async function sendLogin() {
     }).catch((error) => {
         console.error(error);
         if (error.response) {
-            errorMessage.value = `Failed to sign up because of error ${error.response.data}`;
+            errorMessage.value = `Failed to sign in because of error ${error.response.data}`;
         } else {
             errorMessage.value = `Failed to login`;
         }
@@ -71,8 +67,8 @@ form {
     <div class="login-screen">
         <form @submit.prevent="sendLogin">
             <div>
-                <label for="username">Username:</label>
-                <input id="username" type="text" required v-model="authModel.username"/>
+                <label for="email">Email:</label>
+                <input id="email" type="email" required v-model="authModel.email"/>
             </div>
             <div>
                 <label for="password">Password:</label>