+++ src/shared/model.ts
@@ -6,68 +6,13 @@ export interface EchoResponse {
     "string": string;
 }
 
-// AUTH Requests
-export interface GetNonceRequest {
-    "authRequestId": string;
-}
-
-export interface GetNonceResponse {
-    "authRequestId": string;
-    "nonce": string;
-}
-
-export interface CreateAccountRequest {
-    "authRequestId": string;
-    "payload": string;
-}
-
-export interface CreateAccountResponse {
-    "authRequestId": string;
-    "jwt": string;
-}
-
-export interface GetAuthTokenRequest {
-    "authRequestId": string;
-    "payload": string;
-}
-
-export interface GetAuthTokenResponse {
-    "authRequestId": string;
-    "jwt": string;
-}
-
-export interface BatchGetUsernameRequest {
-    "accountIds": string[];
-}
-
-export interface BatchGetUsernameResponse {
-    "accounts": Account[];
-    "failures": AccountFailure[];
-}
-
-export interface Account {
-    "accountId": string;
-    "username": string;
-}
-
-export interface AccountFailure {
-    "accountId": string;
-    "failureReason": string;
-    "failureCode": number;
-}
-
-export interface AccountPayload {
-    "username": string;
-    "password": string;
-}
-
 // GAME Requests
 import type { GameType, GameStatus } from "./engine-types.js";
 export type { GameType, GameStatus };
 
 export interface CreateGameRequest {
     "gameType": GameType;
-    "numPlayers": number;
+    "maxPlayers": number;
     "gameOptions": { [key: string]: string };
 }
 
@@ -86,7 +31,6 @@ export interface JoinGameResponse {
 }
 
 export interface GetGameStateRequest {
-    "accountId": string;
     "gameId": string;
 }
 
@@ -99,7 +43,7 @@ export interface SerializableGame {
     "gameId": string;
     "gameType": GameType;
     "maxPlayers": number;
-    "accountIds": string[];
+    "playerIds": string[];
     "status": GameStatus;
     "state": SerializableGameState;
 }